+++ configs/http/middleware.php
@@ -4,6 +4,8 @@
  *
  * Defines the list of global middleware to be run on all routes
  */
+use Project\HTTP\Middleware\Session;
+
 /**
  * ----------------------------------------------------------
  * List of HTTP-specific middleware
@@ -17,4 +19,5 @@ return [
      *
      * List any HTTP middleware you'd like here
      */
+    Session::class
 ];